Externalize finallyCommitAllChanges and saveToUncommittedChanges
